# packages/api-schemas/back-schemas/openapi.json
{"openapi":"3.0.0","paths":{"/api/osu/beatmapsets/search":{"get":{"operationId":"OsuController_beatmapsetsSearch","parameters":[{"name":"q","required":false,"in":"query","schema":{"maxLength":100,"default":"","type":"string"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetOsuBeatmapsetsSearchResponseDto_Output"}}}}},"tags":["Osu"]}},"/api/scores/{beatmapId}/leaderboard":{"get":{"operationId":"ScoresController_getBeatmapLeaderboard","parameters":[{"name":"beatmapId","required":true,"in":"path","x-nestjs_zod-parent-additional-properties":false,"schema":{"minLength":1,"maxLength":32,"type":"string"}},{"name":"scoreVersion","required":false,"in":"query","x-nestjs_zod-parent-additional-properties":false,"schema":{"minimum":1,"maximum":3,"default":3,"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetScoresBeatmapLeaderboardResponseDto_Output"}}}}},"tags":["Scores"]}},"/api/scores/{beatmapId}/personal-best/{playerId}":{"get":{"operationId":"ScoresController_getBeatmapPersonalBest","parameters":[{"name":"beatmapId","required":true,"in":"path","x-nestjs_zod-parent-additional-properties":false,"schema":{"minLength":1,"maxLength":32,"type":"string"}},{"name":"playerId","required":true,"in":"path","x-nestjs_zod-parent-additional-properties":false,"schema":{"format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$","type":"string"}},{"name":"scoreVersion","required":false,"in":"query","x-nestjs_zod-parent-additional-properties":false,"schema":{"minimum":1,"maximum":3,"default":3,"type":"number"}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetScoresBeatmapPersonalBestResponseDto_Output"}}}}},"tags":["Scores"]}},"/api/scores/submit":{"post":{"operationId":"ScoresController_submitScore","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostScoresSubmitBodyDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostScoresSubmitResponseDto_Output"}}}}},"tags":["Scores"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostAuthLoginBodyDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostAuthLoginResponseDto_Output"}}}}},"tags":["Auth"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostAuthRegisterBodyDto"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostAuthRegisterResponseDto_Output"}}}}},"tags":["Auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetAuthMeResponseDto_Output"}}}}},"tags":["Auth"]}}},"info":{"title":"Tau API","description":"API documentation for the Tau backend","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"GetOsuBeatmapsetsSearchResponseDto_Output":{"type":"object","properties":{"beatmapsets":{"type":"array","items":{"type":"object","properties":{"beatmaps":{"type":"array","items":{"type":"object","properties":{"url":{"type":"string"},"mode":{"type":"string"},"difficulty_rating":{"type":"number"},"id":{"type":"number"}},"required":["url","mode","difficulty_rating","id"],"additionalProperties":false}},"artist":{"type":"string"},"title":{"type":"string"},"creator":{"type":"string"},"id":{"type":"number"},"covers":{"type":"object","propertyNames":{"type":"string"},"additionalProperties":{"type":"string"}}},"required":["beatmaps","artist","title","creator","id","covers"],"additionalProperties":false}}},"required":["beatmapsets"],"additionalProperties":false},"GetScoresBeatmapLeaderboardResponseDto_Output":{"type":"object","properties":{"leaderboard":{"type":"array","items":{"type":"object","properties":{"id":{"type":"integer","minimum":0,"maximum":9007199254740991},"playerName":{"type":"string","minLength":1,"maxLength":32,"nullable":true},"playerId":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$","nullable":true},"beatmapId":{"type":"string","minLength":1,"maxLength":32},"score":{"type":"integer","minimum":0,"maximum":9007199254740991},"maxCombo":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracy":{"type":"number","minimum":0},"missCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"mehCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"goodCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"greatCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"perfectCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"submissionTime":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"scoreVersion":{"type":"integer","minimum":0,"maximum":9007199254740991},"player":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"},"username":{"type":"string","minLength":3,"maxLength":32},"email":{"not":{}},"isEmailVerified":{"not":{}},"hashedPassword":{"not":{}},"totalScore":{"type":"integer","minimum":0,"maximum":9007199254740991},"totalSubmittedScores":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracySum":{"type":"number","minimum":0},"performancePoints":{"type":"number","minimum":0},"ssCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"fullComboCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"createdAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"updatedAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"}},"required":["id","username","totalScore","totalSubmittedScores","accuracySum","performancePoints","ssCount","fullComboCount","createdAt","updatedAt"],"additionalProperties":false,"nullable":true}},"required":["id","playerName","playerId","beatmapId","score","maxCombo","accuracy","missCount","mehCount","goodCount","greatCount","perfectCount","submissionTime","scoreVersion"],"additionalProperties":false}}},"required":["leaderboard"],"additionalProperties":false},"GetScoresBeatmapPersonalBestResponseDto_Output":{"type":"object","properties":{"id":{"type":"integer","minimum":0,"maximum":9007199254740991},"playerName":{"type":"string","minLength":1,"maxLength":32,"nullable":true},"playerId":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$","nullable":true},"beatmapId":{"type":"string","minLength":1,"maxLength":32},"score":{"type":"integer","minimum":0,"maximum":9007199254740991},"maxCombo":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracy":{"type":"number","minimum":0},"missCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"mehCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"goodCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"greatCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"perfectCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"submissionTime":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"scoreVersion":{"type":"integer","minimum":0,"maximum":9007199254740991},"player":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"},"username":{"type":"string","minLength":3,"maxLength":32},"email":{"not":{}},"isEmailVerified":{"not":{}},"hashedPassword":{"not":{}},"totalScore":{"type":"integer","minimum":0,"maximum":9007199254740991},"totalSubmittedScores":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracySum":{"type":"number","minimum":0},"performancePoints":{"type":"number","minimum":0},"ssCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"fullComboCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"createdAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"updatedAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"}},"required":["id","username","totalScore","totalSubmittedScores","accuracySum","performancePoints","ssCount","fullComboCount","createdAt","updatedAt"],"additionalProperties":false,"nullable":true}},"required":["id","playerName","playerId","beatmapId","score","maxCombo","accuracy","missCount","mehCount","goodCount","greatCount","perfectCount","submissionTime","scoreVersion"],"additionalProperties":false},"PostScoresSubmitBodyDto":{"type":"object","properties":{"beatmapId":{"type":"string","minLength":1,"maxLength":32},"score":{"type":"integer","minimum":0,"maximum":9007199254740991},"maxCombo":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracy":{"type":"number","minimum":0},"missCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"mehCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"goodCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"greatCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"perfectCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"player":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"},"username":{"type":"string","minLength":3,"maxLength":32},"email":{"type":"string","format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"},"isEmailVerified":{"type":"boolean"},"hashedPassword":{"type":"string"},"totalScore":{"type":"integer","minimum":0,"maximum":9007199254740991},"totalSubmittedScores":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracySum":{"type":"number","minimum":0},"performancePoints":{"type":"number","minimum":0},"ssCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"fullComboCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"createdAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"updatedAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"}},"required":["id","username","email","isEmailVerified","hashedPassword","totalScore","totalSubmittedScores","accuracySum","performancePoints","ssCount","fullComboCount","createdAt","updatedAt"],"additionalProperties":false,"nullable":true}},"required":["beatmapId","score","maxCombo","accuracy","missCount","mehCount","goodCount","greatCount","perfectCount"],"additionalProperties":false},"PostScoresSubmitResponseDto_Output":{"type":"object","properties":{"wasUploaded":{"type":"boolean"},"score":{"type":"object","properties":{"id":{"type":"integer","minimum":0,"maximum":9007199254740991},"playerName":{"type":"string","minLength":1,"maxLength":32,"nullable":true},"playerId":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$","nullable":true},"beatmapId":{"type":"string","minLength":1,"maxLength":32},"score":{"type":"integer","minimum":0,"maximum":9007199254740991},"maxCombo":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracy":{"type":"number","minimum":0},"missCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"mehCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"goodCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"greatCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"perfectCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"submissionTime":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"scoreVersion":{"type":"integer","minimum":0,"maximum":9007199254740991},"player":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"},"username":{"type":"string","minLength":3,"maxLength":32},"email":{"not":{}},"isEmailVerified":{"not":{}},"hashedPassword":{"not":{}},"totalScore":{"type":"integer","minimum":0,"maximum":9007199254740991},"totalSubmittedScores":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracySum":{"type":"number","minimum":0},"performancePoints":{"type":"number","minimum":0},"ssCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"fullComboCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"createdAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"updatedAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"}},"required":["id","username","totalScore","totalSubmittedScores","accuracySum","performancePoints","ssCount","fullComboCount","createdAt","updatedAt"],"additionalProperties":false,"nullable":true}},"required":["id","playerName","playerId","beatmapId","score","maxCombo","accuracy","missCount","mehCount","goodCount","greatCount","perfectCount","submissionTime","scoreVersion"],"additionalProperties":false}},"required":["wasUploaded","score"],"additionalProperties":false},"PostAuthLoginBodyDto":{"type":"object","properties":{"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["username","password"],"additionalProperties":false},"PostAuthLoginResponseDto_Output":{"type":"object","properties":{"user":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"},"username":{"type":"string","minLength":3,"maxLength":32},"email":{"type":"string","format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"},"isEmailVerified":{"type":"boolean"},"hashedPassword":{"not":{}},"totalScore":{"type":"integer","minimum":0,"maximum":9007199254740991},"totalSubmittedScores":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracySum":{"type":"number","minimum":0},"performancePoints":{"type":"number","minimum":0},"ssCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"fullComboCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"createdAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"updatedAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"}},"required":["id","username","email","isEmailVerified","totalScore","totalSubmittedScores","accuracySum","performancePoints","ssCount","fullComboCount","createdAt","updatedAt"],"additionalProperties":false},"token":{"type":"string","format":"jwt"}},"required":["user","token"],"additionalProperties":false},"PostAuthRegisterBodyDto":{"type":"object","properties":{"username":{"type":"string","minLength":3,"maxLength":30},"email":{"type":"string","format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"},"password":{"type":"string","minLength":8,"maxLength":32}},"required":["username","email","password"],"additionalProperties":false},"PostAuthRegisterResponseDto_Output":{"type":"object","properties":{"user":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"},"username":{"type":"string","minLength":3,"maxLength":32},"email":{"type":"string","format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"},"isEmailVerified":{"type":"boolean"},"hashedPassword":{"not":{}},"totalScore":{"type":"integer","minimum":0,"maximum":9007199254740991},"totalSubmittedScores":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracySum":{"type":"number","minimum":0},"performancePoints":{"type":"number","minimum":0},"ssCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"fullComboCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"createdAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"updatedAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"}},"required":["id","username","email","isEmailVerified","totalScore","totalSubmittedScores","accuracySum","performancePoints","ssCount","fullComboCount","createdAt","updatedAt"],"additionalProperties":false},"token":{"type":"string","format":"jwt"}},"required":["user","token"],"additionalProperties":false},"GetAuthMeResponseDto_Output":{"type":"object","properties":{"user":{"type":"object","properties":{"id":{"type":"string","format":"uuid","pattern":"^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"},"username":{"type":"string","minLength":3,"maxLength":32},"email":{"type":"string","format":"email","pattern":"^(?!\\.)(?!.*\\.\\.)([A-Za-z0-9_'+\\-\\.]*)[A-Za-z0-9_+-]@([A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"},"isEmailVerified":{"type":"boolean"},"hashedPassword":{"not":{}},"totalScore":{"type":"integer","minimum":0,"maximum":9007199254740991},"totalSubmittedScores":{"type":"integer","minimum":0,"maximum":9007199254740991},"accuracySum":{"type":"number","minimum":0},"performancePoints":{"type":"number","minimum":0},"ssCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"fullComboCount":{"type":"integer","minimum":0,"maximum":9007199254740991},"createdAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"},"updatedAt":{"type":"string","format":"date-time","pattern":"^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"}},"required":["id","username","email","isEmailVerified","totalScore","totalSubmittedScores","accuracySum","performancePoints","ssCount","fullComboCount","createdAt","updatedAt"],"additionalProperties":false}},"required":["user"],"additionalProperties":false}}}}